backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.18.4",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.13.0",
    "asyncpg==0.31.0",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "charset-normalizer==3.4.7",
    "click==8.3.2",
    "colorama==0.4.6",
    "confluent-kafka==2.14.0",
    "cryptography==46.0.6",
    "distro==1.9.0",
    "ecdsa==0.19.2",
    "fastapi==0.135.3",
    "google-auth==2.49.1",
    "google-genai==1.70.0",
    "greenlet==3.3.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "mako==1.3.10",
    "markupsafe==3.0.3",
    "mcp>=1.27.0",
    "numpy==2.4.4",
    "openai-agents>=0.13.5",
    "pgvector==0.4.2",
    "pip==25.3",
    "psycopg2-binary>=2.9.11",
    "pyasn1==0.6.3",
    "pyasn1-modules==0.4.2",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "python-dotenv==1.2.2",
    "python-jose==3.5.0",
    "requests==2.33.1",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.49",
    "starlette==1.0.0",
    "tenacity==9.1.4",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uvicorn==0.43.0",
    "websockets==16.0",
]
